% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docluster_GEM}
\alias{docluster_GEM}
\title{Wrap Seurat RNA clustering}
\usage{
docluster_GEM(mtx = bmmc.data$`Gene Expression`, exp = "DN1_BMMC1")
}
\arguments{
\item{mtx}{sparse Matrix of class "dgCMatrix", each row is a gene, each column is a cell,}

\item{exp}{The name of this sample/experiment}
}
\description{
This function allows you to perform standard sc-RNA clustering
}
\examples{
bmmc.data=Read10X(data.dir = "/lab/solexa_weissman/cweng/Projects/MitoTracing_Velocity/SecondaryAnalysis/Donor01_BMMC_1/CellRanger/Donor01_BMMC_1/outs/filtered_feature_bc_matrix")
docluster_GEM(mtx=bmmc.data$`Gene Expression`,exp="DN1_BMMC1")
}
