% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{Multi_Wrapper}
\alias{Multi_Wrapper}
\title{Wrap Seurat Multiomics clustering}
\usage{
Multi_Wrapper(
 
    path = "/lab/solexa_weissman/cweng/Projects/MitoTracing_Velocity/SecondaryAnalysis/Donor01_CD34_1_Multiomekit/CellRanger/Donor01_CD34_1/outs"
)
}
\arguments{
\item{path}{this should be the path to the cell-ranger results XX/outs}
}
\value{
this returns seurat object with both RNA and ATAC
}
\description{
This function allows you to perform standard sc-multiome clustering
}
\examples{
MultiWrapper(path="XX/CellRanger/Donor01_BMMC_1/outs/")
}
