% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MultiWrapper}
\alias{MultiWrapper}
\title{Wrap Seurat Multiomics clustering}
\usage{
MultiWrapper(
 
    path = "/lab/solexa_weissman/cweng/Projects/MitoTracing_Velocity/SecondaryAnalysis/Donor01_CD34_1_Multiomekit/CellRanger/Donor01_CD34_1/outs"
)
}
\arguments{
\item{path}{this should be the path to the cell-ranger results XX/outs}
}
\description{
This function allows you to perform standard sc-multiome clustering
}
\examples{
bmmc.data=Read10X(data.dir = "/lab/solexa_weissman/cweng/Projects/MitoTracing_Velocity/SecondaryAnalysis/Donor01_BMMC_1/CellRanger/Donor01_BMMC_1/outs/filtered_feature_bc_matrix")
docluster_GEM(mtx=bmmc.data$`Gene Expression`,exp="DN1_BMMC1")
}
