% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{Translate_RNA2ATAC}
\alias{Translate_RNA2ATAC}
\title{Function to translate the RNA barcode into ATAC barcode and add a column}
\usage{
Translate_RNA2ATAC(meta = bmmc.filtered@meta.data, RNAclusterPost = "-1")
}
\arguments{
\item{meta}{a dataframe with the row names as the RNA cell barcode usually with the post -1}

\item{RNAclusterPost}{Usually it is -1(default), but can be changed accordingly}
}
\value{
p from ggplot2
}
\description{
This function allows you to input the metadata with row name as cell barcode
}
\examples{
Translate_RNA2ATAC(meta)
}
